# Micro-optimization: use tuples instead of lists for conditions