This should give a small speed boost to some critical code paths.

The benchmark suite says it's not worse... (though it's not very sensitive).

According to some internets, tuple creation in these situations is ~5x faster than list creation: e.g. http://zwmiller.com/blogs/python_data_structure_speed.html

```
       before           after         ratio
     [7c5ee47e]       [58d5aa2f]
     <master>         <use_tuples>
      1.62±0.04μs      1.61±0.08μs     0.99  benchmarks.BenchMessage.time_attribute_access
       8.60±0.3μs       8.71±0.2μs     1.01  benchmarks.BenchMessage.time_attribute_setting
       14.4±0.7μs       13.8±0.3μs     0.96  benchmarks.BenchMessage.time_init_with_values
       14.0±0.5μs       13.3±0.5μs     0.95  benchmarks.BenchMessage.time_instantiation
         370±10μs         344±10μs     0.93  benchmarks.BenchMessage.time_overhead
       22.7±0.3μs         21.1±1μs     0.93  benchmarks.BenchMessage.time_serialize
              424              424     1.00  benchmarks.MemSuite.mem_instance
```