@cetanu 2021-04-06
Nice. I'm a fan of this in general.